Cleans up Http and HttpSuite
* Separates out HttpServerFixtures in preparation for testing body and response encoders
